form tag, positioning, and depth
Made the calculator a form, where the screen is the text input, and got rid of the lorem ipsum. Will see how I will get the buttons to become functional. Calculation part will come after that.

Changed bottom margin of screen to adjust the positioning of buttons below.

Add same depth effect of the buttons to the body.