Display, Error Handling, and Calculations
Using javaScript:
I got the display to work with the buttons by using click events.

I programmed the equals button to evaluate the expression inputted by the user and then disable the field, make a new field, and answer div. It also checks for math errors/invalid input.

I still need to comment a lot of the of the code I wrote, as I was getting confuse with what I wrote as I was I fixing stuff. There are also spacing issues with multiple input fields.